I broke the logic of converting class name to ClassReflection
There was some code which was converting a variable called $class into a new variable called $class but as a ClassReflection object. When I extracted the specification from the caching code, I completely missed the fact that I was now passing a ClassReflection object into places I was expecting a class name, as well as vice versa. Since we can grab the class name from the class reflection object, it's a safe bet to pass that around and grab the name when we need it.