Add minimal coverage for the interface statement service
I haven't come up with a plan to tackle mocking the class reflections, since I think having a service to just convert the class name to a class reflection would be too much. For this reason, I can't test the rest of the file, unless I create fake classes in the test suite to be reflected.