Do not skip annotation classes
I had merged totalwipeout/edpsuperluminal at some point and accidentally kept the requirement for skipping annotated classes, not realizing that it was not part of the edpsuperluminal master branch. I'm removing the specification for the moment, so that the built cache file for this PR matches master, but we can reverse the commit to skip annotation classes in the future.